finalizacao

+++ src/app/components/produto/produto.component.ts
@@ -4,11 +4,12 @@ import {MatTableModule} from '@angular/material/table' ;
 import { ProdutoService } from '../../services/produto.service';
 import { Router } from '@angular/router';
 import { RouterLink } from '@angular/router';
+import { MatIcon } from '@angular/material/icon';
 
 @Component({
   selector: 'app-produto',
   standalone: true,
-  imports: [MatTableModule, RouterLink],
+  imports: [MatTableModule, RouterLink, MatIcon],
   templateUrl: './produto.component.html',
   styleUrl: './produto.component.scss'
 })
@@ -16,7 +17,7 @@ export class ProdutoComponent {
   @Input()produtoData!:ProdutoInterface;
 
   produtos: ProdutoInterface[] = [];
-  displayedColumns: string[] = ['nome', 'quantidade', 'preco'];
+  displayedColumns: string[] = ['nome', 'quantidade', 'preco', 'acao'];
 
  // Adicione a importação do Router no construtor
 constructor(private produtoService: ProdutoService, private router: Router) {}
@@ -33,14 +34,20 @@ navegarParaAdicionarProduto(): void {
   this.router.navigate(['/adicionar-produto']);
 }
 
-// Excluir produto pelo id
-deleteProduto(id: number): void {
-  this.produtoService.deleteProduto(id).then(() => {
-    alert('Produto deletado com sucesso!');
-    this.produtos = this.produtos.filter((produto: { id: number; }) => produto.id !== id);
-  });
+
+  // Excluir produto pelo id
+  deleteProduto(id: number): void {
+    if (confirm('Tem certeza que deseja deletar este produto?')) {
+      this.produtoService.deleteProduto(id).then(() => {
+        alert('Produto deletado com sucesso!');
+        // Atualiza a lista de produtos após a exclusão
+        this.produtos = this.produtos.filter((produto: ProdutoInterface) => produto.id !== id);
+      }).catch((error) => {
+        console.error('Erro ao deletar produto:', error);
+        alert('Ocorreu um erro ao tentar deletar o produto.');
+      });
 }
   
 }
 
-
+}
